Added a post-processing filter function to add do/dont classes to figures
This could be too brittle to be ready for production, but essentially parses
through the text looking for figure with a figcaption that has either a
Do: or Don't as in its string. If the parser finds this then it replaces
the an attribute-less <figcaption> with <figcaption class="do"> or in the case
of a don't class="dont". Also adds a bold tag around the do or don't in the text