Fixed ASTRO-196 greedy figure replacement
The eleventy build file looked for any instance of 'figure' to add the figure
id supporting the direct linking from the compliance page. This was a greedy
grab that took any 'figure' even in instances like 'configure' and replaced it
with the html element code.

Changed the string search to look for '<figure'

Opportunity to improve would be a regex looking specifially for <figure> elements