Updated Recommended: Portion Marking text
Ran into some confusions on the guidance for portion marking. Updated the Recommended Portion Markings section to include this clarification:
"For individual fields, unless the label itself is classified, we recommend an Unclassified Label (U//FOUO) format to show that the data contained by the label is classified without reducing scannability of the label or removing too much visual focus from the data itself. Note that if the label or title of an element IS classified then the portion marking must come before that text like so: (U//FOUO) Classified Label."