Astro 576 add netlify link checker to build process (#136)
* Added initial configuration for link Netlify link checker

This is an out of the box setting for testing purpose.

* Added a todo pattern for anchor links

* Updated todo pattern

* Updated todo pattern

* Add skip pattern

* Updating skip pattern

* Added a blank.png to meet needs of compliance

Added the blank.png asset to accomodate the layout of
Classigication Markings where row 2 has no column. This
should be addressed in a future update removing the need
for a blank.png

Option 1:
Swap the order so the element with no corresponding don't
is on the last row

Option 2:
Review the rest of the site there may be similar orphans
without relying on a blank PNG

Option 3:
Update the two-col layout to support orphaned dos and don'ts

* Added a todo/skip pattern for anchor links

Updating skip pattern

* Added a blank.png to meet needs of compliance

Added the blank.png asset to accomodate the layout of
Classigication Markings where row 2 has no column. This
should be addressed in a future update removing the need
for a blank.png

Option 1:
Swap the order so the element with no corresponding don't
is on the last row

Option 2:
Review the rest of the site there may be similar orphans
without relying on a blank PNG

Option 3:
Update the two-col layout to support orphaned dos and don'ts

* Fixing missing donut png